requirements.txt
opencv-python==4.8.1
numpy==1.26.0
ultralytics==8.0.196
supervision==0.16.0
PyYAML==6.0.1
torch>=2.0.0
torchvision>=0.15.0 